CUR-4667 Rename the lti tool type related functionality.
