+++ app/Http/Controllers/Api/V1/CurrikiInteractiveVideoIntegration/Kaltura/KalturaGeneratedAPIClientController.php
@@ -23,7 +23,7 @@ use App\Repositories\LtiTool\LtiToolSettingInterface;
 use App\Exceptions\GeneralException;
 use App\Http\Requests\V1\LtiTool\KalturaAPISettingRequest;
 use App\Repositories\MediaSources\MediaSourcesInterface;
-use App\Repositories\LtiTool\LTIToolTypeInterface;
+use App\Repositories\LtiTool\LtiToolType\LtiToolTypeInterface;
 
 class KalturaGeneratedAPIClientController extends Controller
 {
@@ -42,10 +42,10 @@ class KalturaGeneratedAPIClientController extends Controller
      * @param KalturaMediaEntryFilter $kMEF
      * @param KalturaFilterPager $kFP
      * @param LtiToolSettingInterface $ltiToolSettingRepository
-     * @param LTIToolTypeInterface $ltiToolTypeRepo
+     * @param LtiToolTypeInterface $ltiToolTypeRepo
      */
     public function __construct(KalturaConfiguration $kC, KalturaClient $kClient, KalturaMediaEntryFilter $kMEF,
-        KalturaFilterPager $kFP, LtiToolSettingInterface $ltiToolSettingRepository, LTIToolTypeInterface $ltiToolTypeRepo
+        KalturaFilterPager $kFP, LtiToolSettingInterface $ltiToolSettingRepository, LtiToolTypeInterface $ltiToolTypeRepo
       )
     {
         $this->kalturaConfiguration = $kC;
