+++ app/Repositories/LtiTool/LtiToolSettingInterface.php
@@ -18,4 +18,14 @@ interface LtiToolSettingInterface extends EloquentRepositoryInterface
      * @throws GeneralException
      */
     public function getRowRecordByOrgAndToolType($orgId, $mediaSourcesId);
+
+    /**
+     * To get row record by org and lti_tool_type_id match
+     *
+     * @param $orgId integer
+     * @param $ltiToolTypeId int
+     * @return object
+     * @throws GeneralException
+     */
+    public function getRowRecordByColumnMatch($orgId, $ltiToolTypeId);
 }
